Correctly count unread chat messages to display accurate notification badges
Update PostgreSQL query to cast count(*) to integer, resolving incorrect unread message counts in chat badges.

+++ apps/backend/api/src/services/chat-service.ts
@@ -365,7 +365,7 @@ export class ChatService {
 
     if (!member || !member.lastReadAt) {
       const messages = await db
-        .select({ count: sql<number>`count(*)` })
+        .select({ count: sql<number>`count(*)::int` })
         .from(chatMessages)
         .where(and(
           eq(chatMessages.channelId, channelId),
@@ -376,7 +376,7 @@ export class ChatService {
     }
 
     const unread = await db
-      .select({ count: sql<number>`count(*)` })
+      .select({ count: sql<number>`count(*)::int` })
       .from(chatMessages)
       .where(and(
         eq(chatMessages.channelId, channelId),
